Fixed reflex dictionary problem
The error message was the following:

Error occured while creating OfflineAnalysis with label hPlusAnalysis
---- FatalRootError BEGIN
Fatal Root Error: @SUB=TStreamerInfo::BuildOld
Cannot convert pair<string,vector<MyJet> >::second from type:vector<MyJet> to type:vector<MyJet>, skip element
---- FatalRootError END
---- Configuration END

The reason was that std::vector<MyJet> was used in the map, but it was
unknown to reflex. Introducing the vector as a dummy in the classes.h
solved the problem.