Removed parentJet pointer from MyVertex
For some reason, with CMSSW_3_X_Y and reading the files over dcap
protocol, TBaskets related to primaryVertex and mcPrimaryVertex were
not transferred correctly. If the parentJet was removed, the files
were read fine. I don't really understand what is going on, but now it
seems to work. The associatedTracks method must then be implemented by
some other way.