Bugfix in getMCParticles()
The tau from muonreplacement was not stored to the MC particle list, because
the code required that a particle must have at least one mother. The direct
decay products of the tau were also not stored, because of my error in the
fix 44281a13d21a8dbf5fa4391e4fc79994000c13cb, which wrongly (implicitly)
required that a particle must also have a grandmother.