Script for creating a multicrab task in a subdirectory
Multicrab uses the section names in multicrab.cfg as the task
directory names for the crab tasks. This is inconvenient if one wants
to submit multiple multicrab tasks from the same directory.

This script solves the problem by creating a subdirectory (wich
(multi)crab like timestamp in the directory name), copying
multicrab.cfg, the crab configuration pointed by multicrab.cfg and the
CMSSW python configuration pointed by the crab configuration to the
subdirectory, and finally executing 'multicrab -create' in the
subdirectory.