Fixed the marker size for data
Apparently the size of marker style 20 does indeed depend on the size
of the TPad (although the documentation claims otherwise). The style 6
is about the same size as the 20 with size 1, and it has absolute
size. Now the marker sizes in plots with and without the ratio pad are
the same :)