Moved application of b tagging scale factor to the bodies of SignalAnalysis and QCDMeasurement
This was needed because b tagging code must be run before MET cut in
order to have the results for SignalAnalysisTree, but the scale factor
should be delivereted to EventWeight only after the b tagging cut. The
scale factor must of course be saved to the SignalAnalysisTree, but
that is already done.