Added support for setting PU-reweighted number of all events for MC samples
This is needed because we have skimmed pattuples, and won't have a
handle on the PU-reweighted number of all events without calculating
it separately (with pileupNtuple_cfg.py etc). Then we also need a
simple-to-use mechanism to update the all event count. This mechanism
is implemented in this commit, and the necessary interface change has
been applied to all plot scripts.

The interface change was necessary because for doing the update we
need the unweighted event count, and for non-skimmed datasets
(e.g. signal) the weighted event count is still valid. Hence we need
to access actually both, which means that user should give only the
top-level counter directory (instead of 'weighted'). Whether or not to
use the weighted counts is controlled by additional boolean flag
argument.