technical&bugfix: Added primary vertex filter step before analysis modules
The bug was that TES systematic variation requires a valid selected
primary vertex. This stems from the need of running our tau ID once
before the variation, and our tau ID needs the selected primary
vertex.

After some discussion with Lauri we came to the conclusion that the
cleanest fix would be to move the primary vertex filtering step to be
run before the analysis modules and systematic variations.

This is implemented such that the filtering is applied to all analyses
using the AnalysisConfiguration module. For other analyses, the
situation should look the same as before.