scripts: Add --tryAgainTimes and --tryAgainSeconds to hplusMultiCrabSubmit.py
Although these didn't solve the issue I wanted to solve. Some times
the submission (especially to nodeslab-0002) fails because the ARC
client is unable to connect to the CE. Unfortunately in this case the
exit code from crab is 0, so the "failure detection" does not think
this case is a failure, and continues to submit next jobs. Anyway,
this feature could still be useful to somebody.